refactor : use ParameterizeTest and ValueSource / add unit feature test
