+++ src/test/java/calculator/domain/input/BasicInputTest.java
@@ -1,8 +1,7 @@
-package calculator.domain;
+package calculator.domain.input;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import calculator.domain.input.BasicInput;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
