+++ src/test/java/calculator/util/InputValidatorTest.java
@@ -2,44 +2,38 @@ package calculator.util;
 
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
 
-import calculator.constants.Constants;
-import calculator.domain.StringCalculator;
+import calculator.constants.InvalidInputConstants;
+import calculator.domain.input.Input;
+import calculator.domain.input.InputTypeSorter;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.ValueSource;
 
 public class InputValidatorTest {
 
-    @Test
-    void 음수_포함() {
+    @ParameterizedTest
+    @ValueSource(strings = {"-1,2,3", "//;\\n1;2;-3", "-1"})
+    void 음수_포함(String input) {
         assertThatIllegalArgumentException()
-                .isThrownBy(() -> StringCalculator.calculate("-1,2,3"))
-                .withMessageMatching(Constants.CONTAINS_NEGATIVE);
+                .isThrownBy(InputTypeSorter.sort(input)::createCalculationInputs)
+                .withMessageMatching(InvalidInputConstants.CONTAINS_NEGATIVE);
     }
 
-    @Test
-    void 기본_구분자_아님() {
+    @ParameterizedTest
+    @ValueSource(strings = {"1.2.3", "//;\\n1;2.3"})
+    void 잘못된_구분자_포함(String input) {
         assertThatIllegalArgumentException()
-                .isThrownBy(() -> StringCalculator.calculate("1.2.3"))
-                .withMessageMatching(Constants.WRONG_DELIMITER);
-    }
-
-    @Test
-    void 정의되지_않은_커스텀_구분자() {
-        assertThatIllegalArgumentException()
-                .isThrownBy(() -> StringCalculator.calculate("//;\\n1;2.3"))
-                .withMessageMatching(Constants.UNDEFINED_CUSTOM);
+                .isThrownBy(InputTypeSorter.sort(input)::createCalculationInputs)
+                .withMessageMatching(InvalidInputConstants.WRONG_DELIMITER);
     }
 
     @Test
     void 커스텀_구분자_없음() {
+        Input customInput = InputTypeSorter.sort("//\\n1;2;3");
+
         assertThatIllegalArgumentException()
-                .isThrownBy(() -> StringCalculator.calculate("//\\n1;2;3"))
-                .withMessageMatching(Constants.INVALID_CUSTOM_FORMAT);
+                .isThrownBy(customInput::createCalculationInputs)
+                .withMessageMatching(InvalidInputConstants.INVALID_CUSTOM_FORMAT);
     }
 
-    @Test
-    void 구분자_하나() {
-        assertThatIllegalArgumentException()
-                .isThrownBy(() -> StringCalculator.calculate(";"))
-                .withMessageMatching(Constants.NOT_CONTAINS_NUMBER);
-    }
 }
